# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  mensaje: string;
}

export interface CrearSesion {
  /**
     * @minLength 1
     * @maxLength 120
     */
  nombre: string;
  /** @maxLength 120 */
  facilitador?: string;
}

export interface Sesion {
  id: string;
  codigo: string;
  nombre: string;
  facilitador?: string | null;
  activa: boolean;
  creadaEn: string;
}

export interface Unirse {
  /**
     * @minLength 8
     * @maxLength 64
     */
  dispositivoId: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  nombre: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  equipoId: string;
}

export interface Participante {
  id: string;
  sesionId: string;
  nombre: string;
  equipoId: string;
  creadoEn: string;
}

export type AvanceEstado = { [key: string]: unknown };

export interface Avance {
  participanteId: string;
  puntos: number;
  misiones: string[];
  logros: string[];
  respondidas: number;
  aciertos: number;
  estado: AvanceEstado;
  actualizadoEn: string;
}

export interface Inscripcion {
  sesion: Sesion;
  participante: Participante;
  /** Avance previo del participante, o null si es su primera vez. */
  avance?: Avance | null;
}

export interface RespuestaEjercicio {
  preguntaId: string;
  opcionId: string;
  correcta: boolean;
}

/**
 * Estado completo del cliente, para restaurarlo en otro dispositivo.
 */
export type GuardarAvanceEstado = { [key: string]: unknown };

export interface GuardarAvance {
  /** @minimum 0 */
  puntos: number;
  misiones: string[];
  logros: string[];
  respuestas: RespuestaEjercicio[];
  /** Estado completo del cliente, para restaurarlo en otro dispositivo. */
  estado: GuardarAvanceEstado;
}

export interface FilaTablero {
  participanteId: string;
  nombre: string;
  equipoId: string;
  puntos: number;
  misiones: string[];
  logros: string[];
  respondidas: number;
  aciertos: number;
  vistoEn: string;
}

/**
 * Conteo de participantes por opción elegida.
 */
export type EstadisticaPreguntaOpciones = {[key: string]: number};

export interface EstadisticaPregunta {
  preguntaId: string;
  respondidas: number;
  correctas: number;
  /** Conteo de participantes por opción elegida. */
  opciones: EstadisticaPreguntaOpciones;
}

export interface ResumenTablero {
  participantes: number;
  activosUltimos5Min: number;
  puntosPromedio: number;
  misionesCompletadas: number;
}

export interface Tablero {
  sesion: Sesion;
  resumen: ResumenTablero;
  filas: FilaTablero[];
  preguntas: EstadisticaPregunta[];
}

/**
 * La solicitud no cumple el contrato
 */
export type SolicitudInvalidaResponse = Error;

/**
 * No encontrado
 */
export type NoEncontradoResponse = Error;