# lib/api-zod/src/generated/types/avance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { AvanceEstado } from './avanceEstado';

export interface Avance {
  participanteId: string;
  puntos: number;
  misiones: string[];
  logros: string[];
  respondidas: number;
  aciertos: number;
  estado: AvanceEstado;
  actualizadoEn: Date;
}