# lib/api-zod/src/generated/types/crearSesion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CrearSesion {
  /**
     * @minLength 1
     * @maxLength 120
     */
  nombre: string;
  /** @maxLength 120 */
  facilitador?: string;
}