# lib/api-zod/src/generated/types/estadisticaPregunta.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { EstadisticaPreguntaOpciones } from './estadisticaPreguntaOpciones';

export interface EstadisticaPregunta {
  preguntaId: string;
  respondidas: number;
  correctas: number;
  /** Conteo de participantes por opción elegida. */
  opciones: EstadisticaPreguntaOpciones;
}