# lib/api-zod/src/generated/types/filaTablero.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface FilaTablero {
  participanteId: string;
  nombre: string;
  equipoId: string;
  puntos: number;
  misiones: string[];
  logros: string[];
  respondidas: number;
  aciertos: number;
  vistoEn: Date;
}