# lib/api-zod/src/generated/types/guardarAvance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { GuardarAvanceEstado } from './guardarAvanceEstado';
import type { RespuestaEjercicio } from './respuestaEjercicio';

export interface GuardarAvance {
  /** @minimum 0 */
  puntos: number;
  misiones: string[];
  logros: string[];
  respuestas: RespuestaEjercicio[];
  /** Estado completo del cliente, para restaurarlo en otro dispositivo. */
  estado: GuardarAvanceEstado;
}