# lib/api-zod/src/generated/types/inscripcion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Avance } from './avance';
import type { Participante } from './participante';
import type { Sesion } from './sesion';

export interface Inscripcion {
  sesion: Sesion;
  participante: Participante;
  /** Avance previo del participante, o null si es su primera vez. */
  avance?: Avance | null;
}