# lib/api-zod/src/generated/types/participante.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Participante {
  id: string;
  sesionId: string;
  nombre: string;
  equipoId: string;
  creadoEn: Date;
}