# lib/api-zod/src/generated/types/resumenTablero.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ResumenTablero {
  participantes: number;
  activosUltimos5Min: number;
  puntosPromedio: number;
  misionesCompletadas: number;
}